requirements.txt
# pip install -r requirements.txt
requests
fake-useragent
beautifulsoup4
schedule
lxml
PyYAML